+++ RingSoft.DataEntryControls.WPF/DropDownEditControls/IntegerEditControl.cs
@@ -90,7 +90,7 @@ namespace RingSoft.DataEntryControls.WPF
             return result;
         }
 
-        protected override void GetMinimumValueProperties(out decimal? value, out decimal? minimumValue)
+        protected override void GetMinimumValueProperties(out double? value, out double? minimumValue)
         {
             value = Value;
             minimumValue = MinimumValue;
